DSEGOG-471 refactor ResizeObserver code to run in the popup window (via script in windowPortal)
